requirements.txt
asgiref==3.7.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
Django==4.2.4
djangorestframework==3.14.0
filelock==3.9.0
fsspec==2023.9.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
nltk==3.8.1
numpy==1.24.1
packaging==23.1
Pillow==9.3.0
psycopg2==2.9.7
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.28.1
safetensors==0.3.3
sqlparse==0.4.4
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.1
transformers==4.33.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
